+++ tsup.config.ts
@@ -2,7 +2,11 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig({
   entry: ['src/index.ts', 'src/bin/gtfs-to-html.ts', 'src/app/index.ts'],
-  dts: true,
+  dts: {
+    compilerOptions: {
+      ignoreDeprecations: '6.0',
+    },
+  },
   clean: true,
   format: ['esm'],
   splitting: false,